USBH: cleanup
